requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-community>=0.0.10
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
pydantic>=2.4.2

# Document processing
pymupdf>=1.22.5
pdfplumber>=0.10.2
python-docx>=0.8.11
beautifulsoup4>=4.12.2
newspaper3k>=0.2.8
requests>=2.31.0

# LLM integration
transformers>=4.35.2
torch>=2.0.1
accelerate>=0.23.0
bitsandbytes>=0.41.1
huggingface-hub>=0.19.4

# API and server
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1